refactor(aiModeration.service): make aiModeration.service return decisions based on ai moderation policies
